+++ dingo-store-raft/src/main/java/io/dingodb/store/raft/config/RaftConfiguration.java
@@ -31,8 +31,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class RaftConfiguration {
-
     private NodeOptions node;
     private String raftPath;
-    private Integer port;
+    private Long port;
 }
